+++ src/curve.jl
@@ -1,7 +1,9 @@
-abstract type CurveMethod end
-
-
 # default = Bootstrap
-function curve(method::T=Bootstrap(),instruments) where {T<:CurveMethod}
+function curve(method::T,instruments) where {T<:CurveMethod}
+    method(instruments)
+end
+
+function curve(instruments) where {T<:CurveMethod}
+    method = Bootstrap()
     method(instruments)
 end
