+++ test/runtests.jl
@@ -3,6 +3,7 @@ using Test
 
 include("instruments.jl")
 include("generic.jl")
+include("Constant.jl")
 include("bootstrap.jl")
 include("RateCombination.jl")
 include("SmithWilson.jl")
